Minor performance optimization and split to new class file and cleanup
